requirements.txt
PyYAML==6.0.2
requests==2.32.3
uuid==1.30
fastapi==0.115.12
uvicorn==0.34.1
websockets==15.0.1